+++ tests/Header/ContentRangeTest.php
@@ -6,36 +6,51 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\ContentRange,
     Header,
-    Header\ContentRangeValue
 };
-use Innmind\Immutable\Sequence;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 class ContentRangeTest extends TestCase
 {
-    public function testInterface()
-    {
-        $h = new ContentRange(
-            $cr = new ContentRangeValue('bytes', 0, 42),
-        );
-
-        $this->assertInstanceOf(Header::class, $h);
-        $this->assertSame('Content-Range', $h->name());
-        $v = $h->values();
-        $this->assertInstanceOf(Sequence::class, $v);
-        $this->assertSame($cr, $v->find(static fn() => true)->match(
-            static fn($first) => $first,
-            static fn() => null,
-        ));
-        $this->assertSame('Content-Range: bytes 0-42/*', $h->toString());
-        $this->assertSame('bytes 0-42/*', $h->range()->toString());
-    }
-
     public function testOf()
     {
         $header = ContentRange::of('bytes', 0, 42);
 
         $this->assertInstanceOf(ContentRange::class, $header);
+        $this->assertInstanceOf(Header::class, $header);
         $this->assertSame('Content-Range: bytes 0-42/*', $header->toString());
+    }
+
+    public function testValids()
+    {
+        $this->assertSame(
+            'Content-Range: resources 0-42/*',
+            ContentRange::of('resources', 0, 42)->toString(),
+        );
+        $this->assertSame(
+            'Content-Range: resources 0-499/1234',
+            ContentRange::of('resources', 0, 499, 1234)->toString(),
+        );
+    }
+
+    #[DataProvider('invalids')]
+    public function testReturnNothingWhenInvalidContentRangeValue($unit, $first, $last, $length)
+    {
+        $this->assertNull(ContentRange::maybe($unit, $first, $last, $length)->match(
+            static fn($header) => $header,
+            static fn() => null,
+        ));
+    }
+
+    public static function invalids()
+    {
+        return [
+            ['', 0, 42, null],
+            ['foo', -1, 42, null],
+            ['foo', 0, -42, null],
+            ['foo', 0, 42, -42],
+            ['foo', 100, 42, 142],
+            ['foo', 100, 142, 42],
+        ];
     }
 }