+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - `Innmind\Http\Header\ContentEncoding` constructor is now private, use `::of()` or `::maybe()` named constructors
 - `Innmind\Http\Header\ContentLength` constructor is now private, use `::of()` or `::maybe()` named constructors
 - `Innmind\Http\Header\ContentLocation` constructor is now private, use `::of()` or `::maybe()` named constructors
+- `Innmind\Http\Header\ContentRange` constructor is now private, use `::of()` or `::maybe()` named constructors
 
 ### Removed
 
@@ -39,6 +40,7 @@
 - `Innmind\Http\Header\AuthorizationValue`
 - `Innmind\Http\Header\ContentEncodingValue`
 - `Innmind\Http\Header\ContentLengthValue`
+- `Innmind\Http\Header\ContentRangeValue`
 
 ### Fixed